profile / new index

+++ app/Http/Controllers/back/countryController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\back;
 use App\comment;
 use App\Country;
 use App\news;
+use App\photo;
 use App\User;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
@@ -35,9 +36,17 @@ class countryController extends Controller
         $country = new country();
         $country->title = $request->title;
         $country->des = $request->des;
+//        $country->littledes = $request->littledes;
         $country->save();
 
-        return redirect('administrator/country',compact('news2','com','user'));
+        $photo = explode(',', $request->input('photo_id')[0]);
+
+        foreach ($photo as $ph){
+        $photos = photo::findOrFail($ph);
+        $photos->country_id = $country->id;
+        $photos->save();}
+
+        return redirect()->route('country.index');
     }
 
 
@@ -51,7 +60,7 @@ class countryController extends Controller
         $news2 = news::all()->count();
         $com = comment::all()->count();
         $user = User::all()->count();
-        $country = country::findorFail($id);
+        $country = country::with('photos')->findorFail($id);
         return view('back.country.edit',compact('country','news2','com','user'));
     }
 
@@ -61,7 +70,15 @@ class countryController extends Controller
         $country = country::findOrFail($id);
         $country->title = $request->title;
         $country->des = $request->des;
+//        $country->littledes = $request->littledes;
         $country->save();
+
+        $photo = explode(',', $request->input('photo_id')[0]);
+        foreach ($photo as $ph){
+            $photos = photo::findOrFail($ph);
+            $photos->country_id = $country->id;
+            $photos->save();}
+
 
         return redirect('administrator/country');
     }
@@ -70,8 +87,23 @@ class countryController extends Controller
     public function destroy($id)
     {
         $country = country::findOrFail($id);
-        $country->delete();
+        if ($country->photo){
+            $photo = photo::findOrFail($country->photo->id);
+            $photo->delete();
+            unlink(getcwd() . $photo->path);
+        }
+        if ($country) {
+            $country->delete();
+        }
 
         return redirect('administrator/country');
+    }
+
+    public function delete($id)
+    {
+        $photo = photo::findOrFail($id);
+        $photo->delete();
+        unlink(getcwd() . $photo->path );
+        return back();
     }
 }
